% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAR.R
\name{AAR}
\alias{AAR}
\title{AAR}
\usage{
AAR(X, Y, a)
}
\arguments{
\item{This}{functions takes input data, labels and scalar a>0.}
}
\value{
predictions, RMSE, MAE, and quantiles of AAR.
}
\description{
AAR
}
\examples{
X<-matrix(rexp(200, rate=.1), ncol=20)
Y<-rnorm(10)
AAR(X,Y,a=1)
}
