% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmarks}
\alias{benchmarks}
\title{benchmarks}
\usage{
benchmarks(X, Y)
}
\arguments{
\item{This}{functions takes input data and labels.}
}
\value{
predictions, RMSE, MAD, and quantiles of lm, ee and lasso.
}
\description{
benchmarks
}
\examples{
benchmarks(X,Y)
}
