% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COIRR.R
\name{COIRR}
\alias{COIRR}
\title{COIRR}
\usage{
COIRR(X, Y, a)
}
\arguments{
\item{This}{functions takes input data, labels and scalar a>0.}
}
\value{
predictions, RMSE, MAE, and quantiles of COIRR.
}
\description{
COIRR
}
\examples{
X<-matrix(rexp(200, rate=.1), ncol=20)
Y<-rnorm(10)
COIRR(X,Y,a=1)
}
