% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cr_colors}
\alias{cr_colors}
\title{Function to extract `cr_color_list` colors as hex codes}
\usage{
cr_colors(...)
}
\arguments{
\item{...}{Character names of CR_color_list colors}
}
\description{
Function to extract `cr_color_list` colors as hex codes
}
