% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_font_test.R
\name{cr_font_test}
\alias{cr_font_test}
\title{Test for Adobe Caslon Pro & Lato import and registration}
\usage{
cr_font_test()
}
\description{
\code{cr_font_test} tests to see if default fonts are imported and registered.
}
\details{
Import and register Adobe Caslon Pro & Lato in R with cr_font_install().
}
