% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_cr_main}
\alias{palette_cr_main}
\alias{palette_cr_diverging}
\alias{palette_cr_sequential}
\title{Connor Rothschild custom color palettes}
\format{An object of class \code{character} of length 6.}
\usage{
palette_cr_main

palette_cr_diverging

palette_cr_sequential
}
\description{
Connor Rothschild custom color palettes
}
\concept{CR palettes}
\keyword{datasets}
