% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_bars.R
\name{fix_bars}
\alias{fix_bars}
\title{Custom \link{ggplot2} theme for personal use by Connor Rothschild}
\usage{
fix_bars(type = "continuous", axis = "y")
}
\arguments{
\item{type}{Data type. Options include "continuous" or "discrete"}

\item{axis}{Axis the data belong to. Options include "x" or "y"}
}
\description{
Drops tick axes and axes labels
}
