% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORR.R
\name{ORR}
\alias{ORR}
\title{ORR}
\usage{
ORR(X, Y, a)
}
\arguments{
\item{This}{functions takes input data, labels and scalar a>0.}
}
\value{
predictions, RMSE, MAE, and quantiles of ORR.
}
\description{
ORR
}
\examples{
X<-matrix(rexp(200, rate=.1), ncol=20)
Y<-rnorm(10)
ORR(X,Y,a=1)
}
