% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSLOG.R
\name{OSLOG}
\alias{OSLOG}
\title{OSLOG}
\usage{
OSLOG(X, Y, a)
}
\arguments{
\item{This}{functions takes input data, labels and scalar a>0 .}
}
\value{
predictions, RMSE, MAE, and quantiles of OSLOG.
}
\description{
OSLOG
}
\examples{
X<-matrix(rexp(200, rate=.1), ncol=20)
Y<-rnorm(10)
OSLOG(X,Y,a=1)
}
