% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{set_color_pal}
\alias{set_color_pal}
\title{Custom [ggplot2] theme for personal use by Connor Rothschild}
\usage{
set_color_pal(palette = "categorical", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in cr_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Set custom color palettes in the `cr` library
Color palettes found on http://colorbrewer2.org and are colorblind safe.
}
