\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,CureRate-method}
\title{Methods for Function \code{show}}
\description{
Method for function \code{show}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "CureRate")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}

