% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cr.R
\name{theme_cr}
\alias{theme_cr}
\title{A print \link{ggplot2} theme formatted in CR style}
\usage{
theme_cr(
  base_size = 11,
  base_family = "Lato",
  base_line_size = 0.5,
  base_rect_size = 0.5
)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family.}

\item{base_line_size}{base line}

\item{base_rect_size}{rectangle sizes}
}
\description{
\code{theme_cr} provides a \link{ggplot2} theme formatted according to the cr
website
}
