% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time}
\alias{time}
\title{time}
\usage{
time(X, Y, a)
}
\arguments{
\item{This}{functions takes input data, labels and scalar a>0.}
}
\value{
time statistics taken for AAR, ORR, OSLOG and COIRR.
}
\description{
time
}
\examples{
time(X,Y,a)
}
