% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneAAR.R
\name{tuneAAR}
\alias{tuneAAR}
\title{tuneAAR}
\usage{
tuneAAR(from, to, by, Xtrain, Ytrain)
}
\arguments{
\item{This}{functions takes scalers; start value, end value, increment and input/output of the training set.}
}
\value{
vector of tuning parameter values, and optimum value.
}
\description{
tuneAAR
}
\examples{
X<-matrix(rexp(200, rate=.1), ncol=20)
Y<-rnorm(10)
data<- cbind(Y,X)
sample <- 1:floor(.25*nrow(data))
train <- data[sample, ]
test  <- data[-sample, ]
Xtrain<- as.matrix(train[,-1])
Ytrain<-scale(as.matrix(train[,1]))
Xtest<- as.matrix(test[,-c1])
Ytest<- as.matrix(test[,1])
tuneAAR(0.1,0.2,0.01,Xtrain,Ytrain)
}
