% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undo_cr_theme.R
\name{undo_cr_theme}
\alias{undo_cr_theme}
\title{Custom \link{ggplot2} theme for personal use by Connor Rothschild}
\usage{
undo_cr_theme()
}
\description{
Resets the \link{ggplot2} theme to its default settings
}
