% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_palette.R
\name{view_palette}
\alias{view_palette}
\title{Custom \link{ggplot2} theme for personal use by Connor Rothschild}
\usage{
view_palette(palette = palette_cr_main)
}
\arguments{
\item{palette}{A \code{palette_cr_*} vector from \code{library(cr)}.
Options are \code{palette_cr_main}, \code{palette_cr_diverging}, \code{palette_cr_politics}, and \code{palette_cr_sequential}}
}
\description{
\code{view_palette} displays the colors and hexadecimal codes for \code{palette_cr_*} vectors.
These palettes were generated on www.coolors.co/.
}
\examples{
view_palette()
view_palette(palette_cr_main)

}
