% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{cr_color_list}
\alias{cr_color_list}
\alias{cr_palettes}
\alias{palette_cr_main}
\alias{palette_cr_diverging}
\alias{palette_cr_sequential}
\title{List of CR colors}
\format{
A \code{list}

A \code{list}

An object of class \code{character} of length 6.

An object of class \code{character} of length 9.

An object of class \code{character} of length 9.
}
\usage{
cr_color_list

cr_palettes

palette_cr_main

palette_cr_diverging

palette_cr_sequential
}
\description{
List of CR colors

List of CR Palettes

Connor Rothschild custom color palettes
}
\concept{CR palettes}
\keyword{datasets}
