% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cr_theme.R
\name{set_cr_theme}
\alias{set_cr_theme}
\title{Custom \link{ggplot2} theme for personal use by Connor Rothschild}
\usage{
set_cr_theme(font = "adobe")
}
\arguments{
\item{font}{The font for plot labels, axes, and titles. Options are "adobe" (Adobe Caslon Pro) and "lato" (Lato).}

\item{style}{The default theme style for the R session. Options are "print" or "map".}
}
\description{
\code{set_cr_theme} provides a \link{ggplot2} theme formatted for use by Connor Rothschild
}
